# api_docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "More-Recipes provides a platform for user to share the awesome and exciting  recipe ideas they have invented or learnt.  Suppose a user comes up with a recipe,  he/she can post it on More-Recipes and  get feedback in form of reviews and votes from other user who explore that recipe. user can also keep a list of their favorite recipes on the application.",
    "version": "v1",
    "title": "More Recipes API",
    "contact": {
      "email": "hamdhalah09@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/swagger-api/swagger-ui/blob/master/LICENSE"
    }
  },
  "basePath": "/api/v1/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/user/{userId}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get the details of a current user",
        "description": "Displays details of current user",
        "operationId": "getCurrentUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-token",
            "in": "header",
            "description": "Authentication Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "401": {
            "description": "Invalid Input",
            "schema": {
              "$ref": "#/definitions/SigninError"
            }
          },
          "404": {
            "description": "NotFound",
            "schema": {
              "$ref": "#/definitions/SigninError"
            }
          },
          "422": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/ValidationError"
            }
          },
          "default": {
            "description": "Unable to update user details"
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update user's profile",
        "description": "A user can update their profile",
        "operationId": "updateUserProfile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-token",
            "in": "header",
            "description": "Authentication Token",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "userId",
            "description": "Edit form",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "Validation exception",
            "schema": {
              "$ref": "#/definitions/SignupValidation"
            }
          },
          "403": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/SignupValidation"
            }
          },
          "default": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/myrecipes": {
      "get": {
        "tags": [
          "/recipes"
        ],
        "summary": "View recipes posted by user",
        "description": "Fetch all recipes posted by a user on the system",
        "operationId": "getUserRecipes",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-token",
            "in": "header",
            "description": "Authentication Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetRecipes"
              }
            }
          },
          "default": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UserResponse": {
      "type": "object",
      "required": [
        "userName",
        "email"
      ],
      "properties": {
        "userName": {
          "type": "string",
          "example": "janeDoe"
        },
        "email": {
          "type": "string",
          "example": "josh@ford.com"
        }
      }
    },
...
    "SignupValidation": {
      "type": "object",
      "required": [
        "error"
      ]
    },
...
    "SigninError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "object",
          "error": "User is not registered!",
          "example": {
            "error": "User is not registered"
          }
        }
      }
    },
...
    "ValidationError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "object",
          "example": {
            "error": "Password is required"
          }
        }
      }
    },
...
    "GetRecipes": {
      "type": "object",
      "required": [
        "recipe"
      ],
      "properties": {
        "page": {
          "type": "integer",
          "example": 1
        },
        "totalCount": {
          "type": "integer",
          "example": 5
        },
        "pageCount": {
          "type": "integer",
          "example": 1
        },
        "pageSize": {
          "type": "integer",
          "example": 5
        },
        "allRecipe": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Recipe"
          }
        }
      }
    },
    "Recipe": {
      "type": "object",
      "required": [
        "name",
        "description",
        "ingredients",
        "direction"
      ],
      "properties": {
        "recipeName": {
          "type": "string",
          "example": "rice"
        },
        "description": {
          "type": "string",
          "example": "cook rice"
        },
        "ingredients": {
          "type": "string",
          "example": "rice, water"
        },
        "preptime": {
          "type": "string",
          "example": "2 hours"
        },
        "image": {
          "type": "string",
          "example": "https://res.cloudinary.com/geek-hijabi/image/upload/v1516821128/food_k1uh5z.jpg"
        },
        "upvotes": {
          "type": "integer",
          "example": 0
        },
        "downvotes": {
          "type": "integer",
          "example": 0
        }
      }
    },
...
  }
}